refactor(chat-sidebar): improve code readability and structure for better maintainability
