+++ messages/ar.json
@@ -260,5 +260,10 @@
         }
       }
     }
+  },
+  "chat": {
+    "users": {
+      "notFound": "لا يوجد اي مستخدمين"
+    }
   }
 }
